requirements.txt
discord.py==2.3.2
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
requests==2.31.0
requests-oauthlib==1.3.1
python-dotenv==1.0.0
cryptography==41.0.7
aiohttp==3.9.1
pydantic==2.5.0
python-multipart==0.0.6
